Move `extensions/pkg/controller/reconciler.go` to `pkg/controllerutils/reconciler`
